fix: handle billing notifications and loader in Settings component
