+++ frontend/pages/community/settings/index.tsx
@@ -1,21 +1,28 @@
 import { Box, Divider } from "@mui/material";
 import { Tabs } from "@rootcodelabs/skapp-ui";
+import { useQueryClient } from "@tanstack/react-query";
 import { type NextPage } from "next";
 import { useRouter } from "next/router";
 import { useEffect, useMemo, useState } from "react";
 
 import { useAuth } from "~community/auth/providers/AuthProvider";
 import ContentLayout from "~community/common/components/templates/ContentLayout/ContentLayout";
+import { OBOARDING_LOGOCOLORLOADER_DURATION } from "~community/common/constants/commonConstants";
 import { appModes } from "~community/common/constants/configs";
+import ROUTES from "~community/common/constants/routes";
+import { SUCCESS } from "~community/common/constants/stringConstants";
 import { GlobalLoginMethod } from "~community/common/enums/CommonEnums";
+import { ToastType } from "~community/common/enums/ComponentEnums";
 import useSessionData from "~community/common/hooks/useSessionData";
 import { useTranslator } from "~community/common/hooks/useTranslator";
+import { useToast } from "~community/common/providers/ToastProvider";
 import {
   AdminTypes,
   EmployeeTypes,
   ManagerTypes
 } from "~community/common/types/AuthTypes";
 import { getSettingsTabs } from "~community/settings/utils/settingsTabsUtil";
+import LogoColorLoader from "~enterprise/common/components/molecules/LogoColorLoader/LogoColorLoader";
 import { useGetEnvironment } from "~enterprise/common/hooks/useGetEnvironment";
 import { useCommonEnterpriseStore } from "~enterprise/common/store/commonStore";
 import { getEnterpriseSettingsTabs } from "~enterprise/settings/utils/settingsTabsUtil";
@@ -24,6 +31,9 @@ const Settings: NextPage = () => {
   const { user } = useAuth();
   const router = useRouter();
   const translateText = useTranslator("settings");
+  const billingTranslateText = useTranslator("settingEnterprise", "billing");
+  const { setToastMessage } = useToast();
+  const queryClient = useQueryClient();
   const environment = useGetEnvironment();
   const isEnterprise = environment === appModes.ENTERPRISE;
   const { isEmployee } = useSessionData();
@@ -62,6 +72,56 @@ const Settings: NextPage = () => {
 
   const [activeTab, setActiveTab] = useState(visibleTabs[0]?.id);
 
+  const [showLoader, setShowLoader] = useState(() => {
+    if (typeof globalThis !== "undefined" && globalThis.location) {
+      const params = new URLSearchParams(globalThis.location.search);
+      return params.get("status") === SUCCESS;
+    }
+    return false;
+  });
+
+  // Handle billing cancel notification
+  useEffect(() => {
+    if (typeof globalThis !== "undefined" && globalThis.location) {
+      const params = new URLSearchParams(globalThis.location.search);
+      if (params.get("status") === "cancel") {
+        setToastMessage({
+          toastType: ToastType.ERROR,
+          title: billingTranslateText(["subscriptionErrorToastTitle"]),
+          description: billingTranslateText([
+            "subscriptionErrorToastDescription"
+          ]),
+          open: true
+        });
+      }
+    }
+  }, [setToastMessage, billingTranslateText]);
+
+  // Handle billing success loader + toast + reset
+  useEffect(() => {
+    if (showLoader) {
+      const params = new URLSearchParams(globalThis.location?.search || "");
+      if (params.get("status") === SUCCESS) {
+        const timer = setTimeout(() => {
+          setShowLoader(false);
+          setToastMessage({
+            toastType: ToastType.SUCCESS,
+            title: billingTranslateText(["subscriptionSuccessToastTitle"]),
+            description: billingTranslateText([
+              "subscriptionSuccessToastDescription"
+            ]),
+            open: true
+          });
+          void queryClient.invalidateQueries();
+          void router.push(`${ROUTES.SETTINGS.BASE}?tab=billing`);
+        }, OBOARDING_LOGOCOLORLOADER_DURATION);
+        return () => clearTimeout(timer);
+      } else {
+        setShowLoader(false);
+      }
+    }
+  }, [showLoader, billingTranslateText, setToastMessage, queryClient, router]);
+
   useEffect(() => {
     if (!router.isReady) return;
     const tabParam = router.query.tab as string | undefined;
@@ -75,6 +135,10 @@ const Settings: NextPage = () => {
     const basePath = router.asPath.split("?")[0];
     globalThis.history.replaceState(null, "", `${basePath}?tab=${id}`);
   };
+
+  if (showLoader) {
+    return <LogoColorLoader />;
+  }
 
   return (
     <ContentLayout
